Cleaner API for training MIL with FastAI
API for training MIL models is now:

```
from slideflow import mil

# Get training & model config
config = mil.build_config(model_name, trainer, **kw)
Project.train_mil(config, ...)
```
where ``model_name`` is the name of an architecture (``clam_mb``, ``clam_sb``, ``mil_fc``, ``marugoto``), and ``trainer`` is either ``'fastai'`` or ``'clam'``.
