+++ slideflow/mil/__init__.py
@@ -1,1 +1,5 @@
-from ._trainers import train_mil, train_clam, legacy_train_clam, train_marugoto
+from .train import train_mil, train_clam, train_fastai
+from .train._legacy import legacy_train_clam
+from ._params import (
+    build_config, TrainerConfig, TrainerConfigFastAI, TrainerConfigCLAM
+)
