+++ slideflow/mil/marugoto/__init__.py
@@ -1,4 +1,2 @@
 from . import data
 from . import model
-from . import train
-from .train import train_mil
